cleaned up the cookie session stuff
Cleaned up the session store to actually work as intended.
The previous way was to return a token cookie, but this was flawed as a token cookie was already being returned with the requests from the session store itself.
So i changed it to use the returned token cookie from session rather then one that is generated